fixed some errors in commandParser, display commands working for the most part

+++ src/methodInterfaces/DisplaySetBackground.java
@@ -11,8 +11,7 @@ public class DisplaySetBackground implements DisplayInterface {
 	@Override
 	public double executeCommand(double[] args, SingleTurtle turtle, Display display,
 			DisplayPreferences displayPreferences, EntryManager colorManager, EntryManager shapeManager) {
-		displayPreferences.setDisplayColor((Color) colorManager.getValue(Double.toString(args[0])));
+		displayPreferences.setDisplayColor((Color) colorManager.getValue(Integer.toString((int)args[0])));
 		return args[0];
 	}
-
 }
